Refactor auth to separate Slack auth from Clack auth
Refactor the auth code so that all the interactions with Slack and
Slack OAuth are in one file in server/auth/slack.ts, separate from the
Clack login cookie handling and such.  The Clack login system now has
an AuthSystem object (which currently has only one function) that
handles the specific auth method, and we can initialize that at server
startup based on the .env.  This should give us the space we need to
put in some other auth for just playing around.
